import Test-Smoke 1.18.09 from CPAN
git-cpan-module:   Test-Smoke
git-cpan-version:  1.18.09
git-cpan-authorid: ABELTJE
git-cpan-file:     authors/id/A/AB/ABELTJE/maintsnap/Test-Smoke-1.18.09.tar.gz

+++ t/syncer_copy.t
@@ -1,7 +1,7 @@
 #! /usr/bin/perl -w
 use strict;
 
-# $Id: syncer_copy.t 235 2003-07-15 14:24:23Z abeltje $
+# $Id: syncer_copy.t 461 2003-10-05 10:33:54Z abeltje $
 
 use FindBin;
 use lib $FindBin::Bin;
@@ -32,9 +32,9 @@ SETUP: {
     local *DIR;
     opendir DIR, 'win32' or plan skip_all => "Cannot opendir 'win32': $!";
     while ( my $file = readdir DIR ) {
-        -f $file or next;
-      File::Copy::copy( File::Spec->catfile( 'win32', $file ),
-                        File::Spec->catfile( $subdir, $file ) );
+        -f File::Spec->catfile('win32', $file ) or next;
+        File::Copy::copy( File::Spec->catfile( 'win32', $file ),
+                          File::Spec->catfile( $subdir, $file ) );
     }
     closedir DIR;
     # Create a '.patch'
@@ -88,7 +88,9 @@ SKIP: {
 }
 
 END { 
-    rmtree( $ddir );
-    rmtree( $cdir );
+    unless ( $ENV{SMOKE_DEBUG} ) {
+      rmtree( $ddir );
+      rmtree( $cdir );
+    }
     chdir File::Spec->updir;
 }
